Included Generic FileSharefor WLS Patching

+++ src/main/scripts/installPatch.sh
@@ -5,7 +5,6 @@ function usage()
 cat << USAGE >&2
 Usage:
     -patchFile          PATCH_FILE      WebLogic Patch File
-    -patchNumber        PATCH_NUMBER    WebLogic Patch Number
     -h|?|--help         HELP            Help/Usage info
 USAGE
 
@@ -19,7 +18,6 @@ function get_param()
         case "$1" in    
          -h |?|--help )        usage ;;
          -patchFile   )        PATCH_FILE=$2 ;;
-         -patchNumber )        PATCH_NUMBER=$2 ;;
                      *)        echo 'invalid arguments specified'
                                usage;;
         esac
@@ -30,7 +28,7 @@ function get_param()
 function validate_input()
 {
     
-    if test $# -ne 4
+    if test $# -ne 2
     then
       usage
       exit 1
@@ -43,28 +41,41 @@ function validate_input()
         exit 1
     fi
 
-    if [ -z "${PATCH_NUMBER}" ];
-    then
-        echo "Patch Number not provided."
-        usage
-        exit 1
-    fi
+}
 
+function runCommandAsOracleUser()
+{
+   cmd="$1"
+   echo "Exec command $cmd"
+
+   runuser -l gnsuryan -c "$cmd" &
+   myPid=$!
+   wait $myPid
+   status="$?"
+   echo -e "\nRETVAL=$status"
+
+}
+
+function getReturnCode()
+{
+  retVal="$1"
+  retCode="$(echo "$retVal"|grep "RETVAL" | cut -d'=' -f2)"
+  echo "$retCode"
 }
 
 function copy_patch()
 {
 
-    ls -l ${WLS_FILE_SHARE_MOUNT}
+    ls -l ${WLS_PATCH_FILE_SHARE_MOUNT}
 
-    if [[ ! -f ${WLS_FILE_SHARE_MOUNT}/${PATCH_FILE} ]];
+    if [[ ! -f ${WLS_PATCH_FILE_SHARE_MOUNT}/${PATCH_FILE} ]];
     then
-        echo "Patch file ${PATCH_FILE} not available on file share mount: ${WLS_FILE_SHARE_MOUNT}"
+        echo "Patch file ${PATCH_FILE} not available on file share mount: ${WLS_PATCH_FILE_SHARE_MOUNT}"
         exit 1
     fi
 
     echo "copying patch from file share to patch home directory..."
-    cp ${WLS_FILE_SHARE_MOUNT}/${PATCH_FILE} ${PATCH_HOME_DIR}/${PATCH_FILE}
+    cp ${WLS_PATCH_FILE_SHARE_MOUNT}/${PATCH_FILE} ${PATCH_HOME_DIR}/${PATCH_FILE}
 }
 
 function setup_patch()
@@ -75,20 +86,27 @@ function setup_patch()
     cleanup_patch
 }
 
-
 function cleanup_patch()
 {
     echo "cleaning up..."
-    rm -rf ${PATCH_HOME_DIR}/*
+
+    if [ ! -z "${PATCH_HOME_DIR}" ];
+    then
+        rm -rf ${PATCH_HOME_DIR}/*
+    fi
+
 }
 trap cleanup_patch EXIT
 
 
 function check_opatch()
 {
-   runuser -l oracle -c '. /u01/app/wls/install/oracle/middleware/oracle_home/wlserver/server/bin/setWLSEnv.sh; /u01/app/wls/install/oracle/middleware/oracle_home/OPatch/opatch lsinventory -jdk ${JAVA_HOME} > /dev/null 2>&1'
+   ret=$(runCommandAsOracleUser '. /u01/app/wls/install/oracle/middleware/oracle_home/wlserver/server/bin/setWLSEnv.sh; /u01/app/wls/install/oracle/middleware/oracle_home/OPatch/opatch lsinventory  > /dev/null 2>&1')
+   echo "$ret"
 
-   if [[ $? != 0 ]];
+   retVal=$(getReturnCode "$ret")
+
+   if [[ "$retVal" != "0" ]];
    then
      echo "opatch command failed. Please set WebLogic Classpath appropriately and try again"
      exit 1
@@ -97,23 +115,43 @@ function check_opatch()
    fi
 }
 
+function getPatchNumber()
+{
+    PATCH_NUMBER=$(runuser -l oracle -c ". /u01/app/wls/install/oracle/middleware/oracle_home/wlserver/server/bin/setWLSEnv.sh > /dev/null 2>&1; /u01/app/wls/install/oracle/middleware/oracle_home/OPatch/opatch query --all ${PATCH_HOME_DIR}/${PATCH_FILE} |grep '<ONEOFF'|grep 'REF_ID'|cut -d'\"' -f2")
+    echo "PATCH_NUMBER: ${PATCH_NUMBER}"
+}
+
 function install_patch()
 {
     unzip -d ${PATCH_HOME_DIR} ${PATCH_HOME_DIR}/${PATCH_FILE}
     chown -R oracle:oracle ${PATCH_HOME_DIR}
 
     echo "Applying Patch..."
-    runuser -l oracle -c '. /u01/app/wls/install/oracle/middleware/oracle_home/wlserver/server/bin/setWLSEnv.sh; cd /u01/app/wls/patches; /u01/app/wls/install/oracle/middleware/oracle_home/OPatch/opatch napply -silent -jdk ${JAVA_HOME}'
+    ret=$(runCommandAsOracleUser '. /u01/app/wls/install/oracle/middleware/oracle_home/wlserver/server/bin/setWLSEnv.sh; cd /u01/app/wls/patches; /u01/app/wls/install/oracle/middleware/oracle_home/OPatch/opatch napply -silent')
+    echo "$ret"
+    
+    retVal=$(getReturnCode "$ret")
+
+   if [[ "$retVal" != "0" ]];
+   then
+     echo "opatch command failed. Please set WebLogic Classpath appropriately and try again"
+     exit 1
+   else
+     echo "opatch command verified successfully."
+   fi
+
 }
 
 function verify_patch()
 {
     echo "Listing all Patches to see if it contains existing patch"
     
-    runuser -l oracle -c '. /u01/app/wls/install/oracle/middleware/oracle_home/wlserver/server/bin/setWLSEnv.sh; /u01/app/wls/install/oracle/middleware/oracle_home/OPatch/opatch lsinventory -jdk ${JAVA_HOME} | grep "Patch  ${PATCH_NUMBER}"'
-    patchApplied=$?
+    ret=$(runCommandAsOracleUser '. /u01/app/wls/install/oracle/middleware/oracle_home/wlserver/server/bin/setWLSEnv.sh; /u01/app/wls/install/oracle/middleware/oracle_home/OPatch/opatch lsinventory -jdk ${JAVA_HOME} | grep "Patch  ${PATCH_NUMBER}"')
+    echo "$ret"
 
-    if [ $patchApplied == "0" ];
+    retVal=$(getReturnCode "$ret")
+
+    if [ "$retVal" == "0" ];
     then
       echo "PATCH INSTALL: SUCCESS"
       exit 0
@@ -127,7 +165,7 @@ function verify_patch()
 
 CURR_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 PATCH_HOME_DIR="/u01/app/wls/patches"
-WLS_FILE_SHARE_MOUNT="/mnt/wlsshare"
+WLS_PATCH_FILE_SHARE_MOUNT="/mnt/wlspatchshare"
 
 get_param "$@"
 
@@ -136,6 +174,8 @@ validate_input "$@"
 setup_patch
 
 copy_patch
+
+getPatchNumber
 
 check_opatch
